backend/requirements.txt
Flask==2.3.3
Flask-MongoEngine==1.0.0
Flask-CORS==4.0.0
mongoengine==0.27.0
python-dotenv==1.0.0
pymongo[srv]==4.6.0
dnspython==2.4.2
Werkzeug==2.3.7